rename to mcp_server_vikingdb_memory
