+++ server/mcp_server_vikingdb_memory/README.md
@@ -48,13 +48,13 @@ Optional environment variables:
 The server can be run with either stdio transport (for MCP integration) or SSE transport:
 
 ```bash
-python -m mcp_server_memory.server --transport stdio
+python -m mcp_server_vikingdb_memory.server --transport stdio
 ```
 
 Or:
 
 ```bash
-python -m mcp_server_memory.server --transport sse
+python -m mcp_server_vikingdb_memory.server --transport sse
 ```
 
 ### Available Tools
@@ -96,7 +96,7 @@ To add this server to your MCP configuration, add the following to your MCP sett
       "command": "uvx",
         "args": [
           "--from",
-          "git+https://github.com/volcengine/mcp-server#subdirectory=server/mcp_server_memory",
+          "git+https://github.com/volcengine/mcp-server#subdirectory=server/mcp_server_vikingdb_memory",
           "mcp-server-memory"
         ],
       "env": {
