+++ server/mcp_server_vikingdb_memory/pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
 ]
 
 [project.scripts]
-mcp-server-memory = "mcp_server_memory.server:main"
+mcp-server-memory = "mcp_server_vikingdb_memory.server:main"
 
 [build-system]
 requires = ["hatchling"]
