+++ server/mcp_server_vikingdb_memory/src/mcp_server_vikingdb_memory/common/auth.py
@@ -5,7 +5,7 @@ import requests
 from volcengine.auth.SignerV4 import SignerV4
 from volcengine.base.Request import Request
 from volcengine.Credentials import Credentials
-from mcp_server_memory.config import config
+from mcp_server_vikingdb_memory.config import config
 
 def prepare_request(method, path, ak, sk, params=None, data=None, doseq=0):
     if params:
